requirements.txt
psutil>=5.9.0
dnspython>=2.4.0
